refactor: rating 서비스 의존 제거

+++ src/main/java/org/movie/reviewer/domain/movie/dto/MovieConverter.java
@@ -29,7 +29,7 @@ public class MovieConverter {
         .build();
   }
 
-  public static MovieTitleResponse toMovieTitleResponse(MovieTitleResponse response, Float rating) {
+  public static MovieTitleResponse toMovieTitleResponse(MovieTitleResponse response, Double rating) {
     return MovieTitleResponse.builder()
         .id(response.getId())
         .title(response.getTitle())
